fourth commit

+++ src/components/Expenses/ExpenseDate.jsx
@@ -2,7 +2,7 @@ import React from "react";
 import "./expensedate.css";
 
 const ExpenseDate = (props) => {
-  const month = props.date.toLocaleString("en-US", { month: "long" });
+  const month = props.date.toLocaleString("en-US", { month: "short" });
   const day = props.date.toLocaleString("en-US", { day: "2-digit" });
   const year = props.date.getFullYear();
   return (
